+++ lib/ts/recipe/session/types.ts
@@ -175,12 +175,12 @@ export interface RecipeInterface {
 
     revokeMultipleSessions(input: { sessionHandles: string[] }): Promise<string[]>;
 
-    /** @deprecated Use getSessionInformation() instead **/
+    /** @deprecated Use getSessionInformation instead **/
     getSessionData(input: { sessionHandle: string }): Promise<any>;
 
     updateSessionData(input: { sessionHandle: string; newSessionData: any }): Promise<void>;
 
-    /** @deprecated Use getSessionInformation() instead **/
+    /** @deprecated Use getSessionInformation instead **/
     getJWTPayload(input: { sessionHandle: string }): Promise<any>;
 
     updateJWTPayload(input: { sessionHandle: string; newJWTPayload: any }): Promise<void>;
@@ -242,8 +242,8 @@ export interface APIInterface {
 export type SessionInformation = {
     sessionHandle: string;
     userId: string;
-    userDataInDatabase: any;
+    sessionData: any;
     expiry: number;
-    userDataInJWT: any;
+    jwtPayload: any;
     timeCreated: number;
 };